+++ src/Api/Configuration/ConfigurationAdapterDecorator.php
@@ -2,12 +2,13 @@
 
 namespace Elio\ElioBatteryIncludedSearchExtension\Api\Configuration;
 
+use Elio\ElioBatteryIncludedSearchExtension\Configuration\BatteryIncludedConfiguration;
 use Elio\ElioDataDiscovery\Api\Configuration\Response\PresetConfigurationResponse;
 use Elio\ElioBatteryIncludedSearchExtension\Api\ApiClientFactory;
-use Elio\ElioBatteryIncludedSearchExtension\Api\Service\LocaleService;
 use Elio\ElioDataDiscovery\Api\Configuration\ConfigurationAdapter;
 use Elio\ElioDataDiscovery\Api\Configuration\Request\ConfigurationRequest;
 use Elio\ElioDataDiscovery\Api\Configuration\Response\ConfigurationResponseCollection;
+use Elio\ElioDataDiscovery\Configuration\ElioDataDiscoveryConfigServiceInterface;
 use Psr\Log\LoggerInterface;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
 
@@ -15,6 +16,7 @@ class ConfigurationAdapterDecorator extends ConfigurationAdapter
 {
     public function __construct(
         private readonly ApiClientFactory $apiClientFactory,
+        private readonly ElioDataDiscoveryConfigServiceInterface $configService,
         LoggerInterface $logger
     )
     {
@@ -23,10 +25,12 @@ class ConfigurationAdapterDecorator extends ConfigurationAdapter
 
     public function getConfig(ConfigurationRequest $request, SalesChannelContext $context): ConfigurationResponseCollection
     {
+        /** @var BatteryIncludedConfiguration $config */
+        $config = $this->configService->getByContext($context)->getExtension(BatteryIncludedConfiguration::NAME);
         $apiClient = $this->apiClientFactory->createSearchApi($context);
         $presets = $apiClient->configuration($request->getType());
         $response = new ConfigurationResponseCollection();
-        $response->addConfigurationResponse(new PresetConfigurationResponse($presets));
+        $response->addConfigurationResponse(new PresetConfigurationResponse($presets, $config->getCollection()));
         return $response;
     }
 }