+++ src/Api/Search/ResponseTransformer/SuggestProductTransformer.php
@@ -5,27 +5,14 @@ namespace Elio\ElioBatteryIncludedSearchExtension\Api\Search\ResponseTransformer
 use Elio\ElioBatteryIncludedApiClient\Model\SuggestionResultCollection;
 use Elio\ElioDataDiscovery\Api\Request\ApiRequest;
 use Elio\ElioDataDiscovery\Api\Response\ResponseCollection;
-use Elio\ElioDataDiscovery\Api\Search\Components\SuggestTypes;
-use Elio\ElioDataDiscovery\Api\Search\Response\SuggestionResponse;
-use Elio\ElioDataDiscovery\Api\Transform\ResponseTransformerInterface;
-use Elio\ElioDataDiscovery\Configuration\ElioDataDiscoveryConfigServiceInterface;
+use Elio\ElioDataDiscovery\Api\Search\ResponseTransformer\AbstractSuggestProductTransformer;
 use Elio\ElioDataDiscovery\Core\Exception\InvalidTypeException;
-use Elio\ElioDataDiscovery\Core\Suggest\SuggestGroup;
 use Elio\ElioDataDiscovery\Core\Suggest\SuggestItem;
 use Elio\ElioDataDiscovery\Swagger\ModelInterface;
-use Shopware\Core\Content\Product\ProductEntity;
-use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
-use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsAnyFilter;
-use Shopware\Core\System\SalesChannel\Entity\SalesChannelRepository;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
 
-class SuggestProductTransformer implements ResponseTransformerInterface
+class SuggestProductTransformer extends AbstractSuggestProductTransformer
 {
-    public function __construct(
-        private readonly SalesChannelRepository $productRepository,
-        private readonly ElioDataDiscoveryConfigServiceInterface $configService
-    ) {}
-
     public function supports(ModelInterface $model, ApiRequest $request, SalesChannelContext $context): bool
     {
         return $model instanceof SuggestionResultCollection;
@@ -37,70 +24,21 @@ class SuggestProductTransformer implements ResponseTransformerInterface
             throw new InvalidTypeException($model, SuggestionResultCollection::class);
         }
 
-        /** @var SuggestionResponse|null $suggestionResponse */
-        $suggestionResponse = $responseCollection->get(SuggestionResponse::class) ?? new SuggestionResponse();
-        $responseCollection->set(SuggestionResponse::class, $suggestionResponse);
-        $config = $this->configService->getByContext($context);
-        $groupLabels = $config->getSuggestTypeLabels();
-
-        $productGroupKey = SuggestTypes::PRODUCT->value;
-        if (isset($groupLabels[SuggestTypes::PRODUCT->value])) {
-            $productGroupKey = $groupLabels[SuggestTypes::PRODUCT->value];
-        }
-        
-        if(
-            !$suggestionResponse ||
-            !$suggestionResponse->hasGroup($productGroupKey)) {
-            return;
-        }
-
-        $productGroup = $suggestionResponse->getGroup($productGroupKey);
-        $products = $this->collect($productGroup, $context);
-        $this->enrich($productGroup, $products);
-    }
-
-    protected function collect(SuggestGroup $group, SalesChannelContext $context): array
-    {
-        $productNumbers = [];
-        foreach ($group->getItems() as $item) {
-            if($productNumber = $this->getProductNumber($item)) {
-                $productNumbers[] = $productNumber;
-            }
-        }
-
-        if(empty($productNumbers)) {
-            return [];
-        }
-
-        $criteria = new Criteria();
-        $criteria->addFilter(new EqualsAnyFilter('productNumber', $productNumbers));
-        $products = [];
-
-        /** @var ProductEntity $product */
-        foreach ($this->productRepository->search($criteria, $context) as $product) {
-            $products[$product->getProductNumber()] = $product;
-        }
-
-        return $products;
-    }
-
-    protected function enrich(SuggestGroup $group, array $products): void
-    {
-        foreach ($group->getItems() as $item) {
-            $productNumber = $this->getProductNumber($item);
-            if($productNumber && isset($products[$productNumber])) {
-                $item->setEntity($products[$productNumber]);
-            }
-        }
+        parent::transform($model, $responseCollection, $context, $request);
     }
 
     protected function getProductNumber(SuggestItem $item): ?string
     {
         $attributes = $item->getAttributes();
         if (!empty($attributes['ProductNumber'])) {
-            return array_shift($attributes['ProductNumber']);
+            return str_replace(['<mark>', '</mark>'], ['', ''], array_shift($attributes['ProductNumber']));
         }
 
-        return $attributes['MasterProductNumber'] ?? null;
+        $masterProductNumber = $attributes['MasterProductNumber'] ?? null;
+        if ($masterProductNumber) {
+            return str_replace(['<mark>', '</mark>'], ['', ''], $masterProductNumber);
+        }
+
+        return null;
     }
 }