+++ src/Api/Search/ResponseTransformer/Util/LocaleUtil.php
@@ -33,6 +33,9 @@
 namespace Elio\ElioBatteryIncludedSearchExtension\Api\Search\ResponseTransformer\Util;
 
 
+use JetBrains\PhpStorm\Deprecated;
+use Shopware\Core\System\Language\LanguageEntity;
+
 /**
  * Class LocaleFilterUtil
  * @package Api\Search\ResponseTransformer\Util
@@ -45,7 +48,9 @@ class LocaleUtil
 {
     public static function fieldByLocalAllowed(string $fieldName, string $locale): bool
     {
-        return !str_contains($fieldName, '_i18n') || str_contains($fieldName, '_i18n.'.$locale);
+        $pattern = strlen($locale) === 2 ? '/_i18n\.\w{2}\./' : '/_i18n\.[a-z]{2,3}-[A-Z]{2}\./';
+
+        return !preg_match($pattern, $fieldName) || str_contains($fieldName, '_i18n.' . $locale);
     }
 
     public static function normalizeToLocalePlaceholder(string $fieldName, string $locale): string
@@ -56,5 +61,19 @@ class LocaleUtil
     public static function replaceLocalPlaceholder(string $fieldName, string $locale): string
     {
         return str_replace('i18n.%locale%', $locale, $fieldName);
+    }
+
+    public static function getLocaleByLanguage(
+        LanguageEntity $language,
+        #[Deprecated(reason: 'Use full locale code instead', since: '6.6.11')] bool $useLegacyLocale
+    ): string
+    {
+        $locale = $language->getLocale()?->getCode() ?? '--';
+
+        if ($useLegacyLocale) {
+            return substr($locale, 0, 2);
+        }
+
+        return $locale;
     }
 }