+++ src/Configuration/BatteryIncludedConfiguration.php
@@ -32,6 +32,7 @@
 
 namespace Elio\ElioBatteryIncludedSearchExtension\Configuration;
 
+use JetBrains\PhpStorm\Deprecated;
 use Shopware\Core\Framework\Struct\Struct;
 
 /**
@@ -56,7 +57,8 @@ class BatteryIncludedConfiguration extends Struct
         private readonly int $apiTimeOut,
         private readonly string $promotionTemplate,
         private readonly string $suggestPromotionTemplate,
-        private readonly int $navigationStartLevelExport
+        private readonly bool $ignoreLocaleForListingRequest,
+        #[Deprecated(reason: 'Use full locale code instead', since: '6.6.11')] private readonly bool $useLegacyLocale
     ) {}
 
     /**
@@ -119,11 +121,13 @@ class BatteryIncludedConfiguration extends Struct
         return $this->suggestPromotionTemplate;
     }
 
-    /**
-     * @return int
-     */
-    public function getNavigationStartLevelExport(): int
+    public function isIgnoreLocaleForListingRequest(): bool
     {
-        return $this->navigationStartLevelExport;
+        return $this->ignoreLocaleForListingRequest;
+    }
+
+    public function isUseLegacyLocale(): bool
+    {
+        return $this->useLegacyLocale;
     }
 }